Added stack navigator to home screen
