+++ GopherTunnel/styles/base.ts
@@ -1,10 +1,10 @@
-import {StyleSheet} from 'react-native';
+import { StyleSheet } from 'react-native';
 
 export default StyleSheet.create({
   centered: {
     flex: 1,
     flexDirection: 'column',
-    alignItems: 'center',
+    alignItems: 'center'
   },
   routeButton: {
     flex: 1,
@@ -17,5 +17,9 @@ export default StyleSheet.create({
     flexDirection: 'column',
     justifyContent: 'center',
     paddingHorizontal: 10
+  },
+  header: {
+    paddingBottom: 10, 
+    paddingLeft: 5
   }
 });
